requirements.txt
asgiref==3.6.0
attrs==23.1.0
bcrypt==4.0.1
cached-property==1.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==40.0.2
distro==1.8.0
Django==3.2.19
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-extra-fields-3.4.1
djangorestframework-3.14.0
django-filter-23.2
djoser-2.2.0
idna==3.4
importlib-metadata==6.6.0
jsonschema==3.2.0
packaging==23.1
paramiko==3.1.0
pillow-9.5.0
psycopg2-2.9.6
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dotenv==0.21.1
pytz==2023.3
pywin32==306
PyYAML==5.4.1
requests==2.29.0
six==1.16.0
sqlparse==0.4.4
texttable==1.6.7
typing_extensions==4.5.0
urllib3==1.26.15
websocket-client==0.59.0
wheel-0.40.0
zipp==3.15.0
